add context, parallelised, decoupled instruct prompt, model generalisation, folder output
