+++ humun_benchmark/__init__.py
@@ -1,2 +1,8 @@
-from humun_benchmark.models import Model
-from humun_benchmark.prompts import InstructPrompt
+from humun_benchmark.models import HuggingFace, StatisticalModel, MLModel
+
+MODEL_REGISTRY = {
+    "llm": HuggingFace,
+    "my_arima": StatisticalModel,
+    "my_xgboost": MLModel,
+    # …
+}
