--- humun_benchmark/models/__init__.py
@@ -1,2 +1,0 @@
-from .base import Model, ModelLoadError
-from .huggingface import HuggingFace
