+++ humun_benchmark/models/llm/__init__.py
@@ -1,0 +1,2 @@
+from .prompts import Prompt, InstructPrompt
+from .huggingface import HuggingFace
