--- humun_benchmark/prompts/__init__.py
@@ -1,1 +1,0 @@
-from .prompt import Prompt, InstructPrompt
